+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Release 4.0.0:
  - Add `PresentationFactory`: Extracted presentation response building logic from `OidcSiopWallet` and put it here
    - Also added some code for presentation submission validation
  - Update implementation of OpenID 4 Verifiable Credential Issuance, draft 13
+ - Replace `createCredentialRequestJwt()` and `createCredentialRequestCwt()` with `createCredentialRequest()` in `WalletService` for OID4VCI
 
 Release 3.8.0:
  - Kotlin 2.0.0